Relations added between user and listing

+++ routes/web.php
@@ -31,34 +31,34 @@ use App\Http\Controllers\ListingController;
 Route::get("/", [ListingController::class,"index"]);
 
 //show create Form
-Route::get("/listings/create", [ListingController::class,"create"]);
+Route::get("/listings/create", [ListingController::class,"create"])->middleware("auth");
 
 //Store Listing Data
-Route::post("/listings",[ListingController::class,"store"]);
+Route::post("/listings",[ListingController::class,"store"])->middleware("auth");;
 
 //Show edit form
-Route::get("/listings/{listing}/edit",[ListingController::class,"edit"]);
+Route::get("/listings/{listing}/edit",[ListingController::class,"edit"])->middleware("auth");;
 
 //Update Listing
-Route::put("/listings/{listing}",[ListingController::class, "update"]);
+Route::put("/listings/{listing}",[ListingController::class, "update"])->middleware("auth");;
 
 //Delete Listing
-Route::delete("/listings/{listing}",[ListingController::class, "destroy"]);
+Route::delete("/listings/{listing}",[ListingController::class, "destroy"])->middleware("auth");;
 
 //Single Listing
 Route::get("/listings/{listing}",[ListingController::class,"show"]);
 
 //Show register create form
-Route::get("/register",[UserController::class,"create"]);
+Route::get("/register",[UserController::class,"create"])->middleware("guest");
 
 //Create new User
 Route::post("/users", [UserController::class, "store"]);
 
 //Logout
-Route::post("/logout",[UserController::class, "logout"]);
+Route::post("/logout",[UserController::class, "logout"])->middleware("auth");;
 
 //Show login Form
-Route::get("/login",[UserController::class, "login"]);
+Route::get("/login",[UserController::class, "login"])->name("login")->middleware("guest");
 
 //Log in user
 Route::post("/users/authenticate",[UserController::class,"authenticate"]);
